requirements.txt
streamlit>=1.32.0
duckdb>=0.10.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.18.0
pydeck>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
ollama>=0.1.7
ibm-watsonx-ai>=1.0.0
pydantic>=2.5.0
sqlparse>=0.4.4
openpyxl>=3.1.2
xlrd>=2.0.1
rapidfuzz>=3.6.0
streamlit-folium>=0.18.0
folium>=0.15.0
geopy>=2.4.0
